refactor: 실행 메서드 기능별로 메서드 추출
